Add the "tensorflow-model-analysis" package to the notebook images (#544)
* Add the "tensorflow-model-analysis" package to the notebook images

More information about this package is [here](https://github.com/tensorflow/model-analysis).

* Make the tensorflow library importable when building the GPU image.

Installing the `tensorflow_model_analysis` nbextension results in
the `tensorflow` package being imported. For the CPU image this is
fine, but in the GPU image tensorflow relies on `libcuda`, which
is not installed (it is expected to be mapped into the running
container from the host OS).

To work around this, we take the `libcuda.so` stub library (which
is meant for building things that link against libcuda), and
temporarily make it be loaded when something tries to load
`libcuda.so.1`.

* Document the LD_LIBRARY_PATH workaround used for installing the tensorflow-model-analysis extension