Jupyter UI that manages Notebook CRs (#2357)
* Introduce jupyter-web-app's code under /components

Add the webapp's code to /components/jupyter-web-app. It is a
basic Flask application with HTML/CSS/jQuery.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Add jupyter-web-app component in jupyter pkg

Add the jsonnet files in /kubeflow/jupyter. Currently the webapp
is used alongside JupyterHub. The webapp is under the prefix /jupyter.
The image param of the jupyter-web-app must be configured before
deploying.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* UI fixes

Fixes noted from avdaredevil:

* Fix XSS vulnerability
* Use template literals for HTML inside JS

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Add webapp to init scripts

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>