jupyter-web-app: Use SubjectAccessReviews instead of KFAM (#4556)
* Introduce the needs_authorization decorator

Instead of having an if-then block on each of the k8s api calls we will
wrap them on a decorator that will be making this check.

The decorator will need to load the value of the namespace from the
arguments of the decorated function. For this, when we will need to call
one of these api functions we will need to use the namespace as a named
argument in order for the decorator to pick it up.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Use SubjecAccessReviews for authorization

Instead of talking to KFAM use SubjectAccessReview objects.

Should we still allow requests without user header to continue without
applying authorization in order to be compatible with 0.5?

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Rename the api functions to reflect k8s verbs

Replace api.post_ aith api._create as well as use the api.list_ instead
of api.get_ to make it clear which k8s verb the function will be using.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Set the arguments for api calls to correct order

Since the namespace argument will be named we will need to put it in the
end of the arguments in the functions definitions.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>