Jupyter-Web-App: Support for selecting GPUs (#4614)
* Add prettierc file to the frontend

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Add a dev mode in the backend

While developing we want the backend to skip all authorization checks.
We introduce a settings.py module that will contain global variables for
configuring the Backend's behavior,

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* HTML/CSS for GPUs Section

* Create a new Angular Component
* Add the HTML/CSS to have a mockup of how it will look

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* TypeScript logic and the UI mechanics

* The frontend should be able to put a default value for gpus
even if JWA's config is old and does not have the necessary entry
* If the user selects `None` in the number of GPUs, then the UI should
disable the vendor field
* The frontend will send an extra `gpus` field on the POST request to
the backend

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Backend logic for using GPUs

* Backend will work even if the 'gpus' configuration is missing from the
config file

* It will set the limits field of the Notebook Container  with the
appropriate annotation "<vendor>.com/gpu".

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Update the config.yaml

Add a gpus field in the default config. Will need to also update the
manifests repo.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Make the vendor list fully configurable

Add a field in jwa's config to allow the admin to specify which vendors
should be shown in the UI as well as the limits key that correspond to
each one of them.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Fix typo

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>