[Jupyter-Web-App]:Add basic E2E Tests (#4773)
* Add basic e2e tests for JWA

This test has two basic test cases for now.

* Navigate to index page and ensure that no error appeared
* Navigate to form page and ensure that no error appeared

Needs to add logic for IAP authentication

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* Handle IAP login logic

We need to set the driver request headers in order to allow the tests to
conect to the IAP protected cluster. For this we use the selenium-wire
which will create a proxy server which will be setting the necessary
headers.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>