twa: Update the build process and instructions for the Tensorboards web app (#5498)
* web-apps(back): Helper config functions

Introduce helper function for creating the config object for an app.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* twa: Update the build process

Update both the frontend and the backend of the Tensorboards web app to
follow the build/run process of the other web apps as well.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* twa: Update the README

Restructure the README to look like the JWA one. Also update the
instructions with the latest process for running the web app.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>