web-apps: Fixes to the common code (#5567)
* web-apps(back): Fetch the correct default SC

The backend would only check for the annotation of the default
StorageClass but not if it's value would be true/false.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* web-apps(back): Fetch Pod logs

Extend the common backend libraries to fetch pod logs

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* web-apps(front): Add global variables.scss

Since we will need the page padding in multiple places we'll create a
scss file to store these values.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* web-apps(front): Export the Dialog module

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* web-apps(front): Rework the Toolbar component

* Add option to make button to be stroked
* Don't emit an event when a button is clicked. Expect a function which
  will be executed from this component.
* Put all the buttons on the right to mimic Pipelines UI

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* web-apps(front): Make the app's background white

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* web-apps(front): Export commonly used modules

The mat-divider and mat-icon modules are oftenly used so we could
include them in the exports of the KubeflowModule.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* web-apps(front): Key-val should be on top in list

In the list that shows key-valu pairs in lines we want the key to remain
on top if the value has a big height. Previously the key-title would be
in the middle height of the value, which looked weird.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* web-apps(front): Add a Comment component for forms

This is used to explain different sections in the form.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* web-apps(front): Bar at the bottom for every form

We want our forms to have a bar on the bottom with CREATE CANCEL buttons
and the ability to view the yaml contents of a CR.

This component is only used for visualization. The logic component
should be handling this component via this component's inputs/outputs

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* web-apps(front): General purpose css

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* web-apps(front): Make k8s fields optional

We might need to instantiate a k8s objecti with only some of the
subfields, like only the spec and not metadata.

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

* remove redundant link class

Co-authored-by: Tasos Alexiou <tasos@arrikto.com>

* support fit-content for firefox as well

Co-authored-by: Tasos Alexiou <tasos@arrikto.com>

* firefox fit-content support

Co-authored-by: Tasos Alexiou <tasos@arrikto.com>

* use button instead of span

Signed-off-by: Kimonas Sotirchos <kimwnasptd@arrikto.com>

Co-authored-by: Tasos Alexiou <tasos@arrikto.com>