jwa(front): Fix the workspace volume form's inputs (#7029)
* jwa(front): Fix the size input value

The size form input was not showing the correct value, once the frontend
got the ConfigMap defaults. We should be updating the intermediate FormControl
when the data from the ConfigMap arrives at the frontend.

Signed-off-by: Elena Zioga <elena@arrikto.com>

* jwa(front): Fix the access-mode input value

The access mode form input was not showing the correct value, once the frontend
got the ConfigMap defaults. We should be updating the intermediate FormControl
when the data from the ConfigMap arrives at the frontend.

Signed-off-by: Elena Zioga <elena@arrikto.com>

* jwa(front): Fix the name input value

The name form input was not showing the correct value, once the frontend
got the ConfigMap defaults. We should be updating the intermediate FormControl
when the data from the ConfigMap arrives at the frontend.

Signed-off-by: Elena Zioga <elena@arrikto.com>

* jwa(front): Use 5Gi everywhere

All new volumes will have a default value of 5Gi. This includes the 'Add
new volume' button for both the workspace and data volumes.

Signed-off-by: Elena Zioga <elena@arrikto.com>

* jwa(front): Add UI tests with Cypress

Add integration tests with Cypress to ensure that the form will have the
correct values once it gets the ConfigMap.

Signed-off-by: Elena Zioga <elena@arrikto.com>

---------

Signed-off-by: Elena Zioga <elena@arrikto.com>