fix: set default GPU vendors list (#7320)
The change introduced by kubeflow/kubeflow#6736 removed the default GPU vendors
list, which causes an issue when trying to select a vendor from the dropdown menu
if the vendors list is not configured.
This commit can be reverted if proper documentation is provided for users/distributions
to configure the dropdown menu.
Fixes #7273